Retrieving XML script
Working with Xml is strange. If you try and pull the data with standard python you return data but it seems to be memory values and not the data itself. You have to use a module to pull the data correctly. The people on IRC recommend LXML. The course uses a  module called ElementTree.